+++ lib/concorde.bash
@@ -1,7 +1,9 @@
-declare -Ag __feature_hsh
-[[ -n ${__feature_hsh[concorde]:-} && ${1:-} != 'reload' ]] && return
+declare -Ag __features
+[[ -n ${__features[concorde]:-} && ${1:-} != 'reload' ]] && return
 [[ ${1:-} == 'reload' ]] && shift
-__feature_hsh[concorde]="( [root]=$(readlink -f "$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")"/..) )"
+declare -Ag __macros
+type -P greadlink >/dev/null 2>&1 && __macros[readlink]='greadlink -f --' || __macros[readlink]='readlink -f --'
+__features[concorde]="( [root]=$(${__macros[readlink]} "$(dirname "$(${__macros[readlink]} "${BASH_SOURCE[0]}")")"/..) )"
 
 unset -v CDPATH
 
@@ -27,16 +29,16 @@ assign () {
 }
 
 bring () { (
-  [[ $2 == 'from'   ]] || return
-  [[ $1 == '('*')'  ]] && eval "local -a function_ary=$1" || local -a function_ary=( "$1" )
+  [[ $2 == 'from' ]]  || return
+  is_literal "$1"     && eval "local -a function_ary=$1" || local -a function_ary=( "$1" )
   local spec=$3
   local feature
 
   $(require "$spec")
   feature=${spec##*/}
   feature=${feature%.*}
-  $(grab dependencies from_feature "$feature")
-  [[ -n $dependencies ]] && {
+  is_feature "$feature" && $(grab dependencies from_feature "$feature")
+  [[ -n ${dependencies:-} ]] && {
     $(local_ary dependency_ary=$dependencies)
     function_ary+=( "${dependency_ary[@]}" )
   }
@@ -45,9 +47,15 @@ bring () { (
   emit "$__"
 ) }
 
-die () { [[ -n ${1:-} ]] && puterr "$1"; exit "${2:-1}" ;}
+die () {
+  local rc=$?
 
-emit () { printf 'eval eval %q\n' "$1" ;}
+  [[ -n ${1:-} ]] && puterr "$1"
+  exit "${2:-$rc}"
+}
+
+emit          () { printf 'eval eval %q\n' "$1"         ;}
+escape_items  () { printf -v __ '%q ' "$@"; __=${__% }  ;}
 
 _extract_function () {
   local function=$1
@@ -95,7 +103,7 @@ get_str () { IFS=$'\n' read -rd '' __ ||:         ;}
 
 grab () {
   [[ $2 == 'from_feature' || $2 == 'from' ]] || return
-  [[ $2 == 'from_feature' ]] && $(local_hsh arg_hsh=__feature_hsh[$3]) || $(local_hsh arg_hsh=$3)
+  [[ $2 == 'from_feature' ]] && $(local_hsh arg_hsh=__features[$3]) || $(local_hsh arg_hsh=$3)
   case $1 in
     '('*')' ) eval "local -a var_ary=$1"            ;;
     '*'     ) local -a var_ary=( "${!arg_hsh[@]}" ) ;;
@@ -132,7 +140,10 @@ in_scope () {
   emit "$__"
 }
 
-instantiate () { printf -v "$1" %s "$(eval "echo ${!1}")" ;}
+instantiate   () { printf -v "$1" %s "$(eval "echo ${!1}")" ;}
+is_feature    () { is_set __features["$1"]                  ;}
+is_identifier () { [[ $1 =~ ^[_[:alpha:]][_[:alnum:]]*$ ]]  ;}
+is_literal    () { [[ $1 == '('*')' ]] ;}
 
 is_set () {
   set -- "$1" "${1%%[*}"
@@ -149,10 +160,10 @@ feature () {
   local statement
 
   get_here_str <<'  EOS'
-    declare -Ag __feature_hsh
-    [[ -n ${__feature_hsh[%s]:-} && $1 != 'reload' ]] && return
-    [[ ${1:-} == 'reload' ]]  && shift
-    __feature_hsh[%s]="( [root]=$(readlink -f "$(dirname "$(readlink -f "${BASH_SOURCE[0]}")")"%s) )"
+    declare -Ag __features
+    [[ -n ${__features[%s]:-} && ${1:-} != 'reload' ]] && return
+    [[ ${1:-} == 'reload' ]] && shift
+    __features[%s]="( [root]=$(${__macros[readlink]} "$(dirname "$(${__macros[readlink]} "${BASH_SOURCE[0]}")")"%s) )"
   EOS
   statement=$__
   path=''
@@ -170,30 +181,37 @@ local_ary () {
 
   name=${first%%=*}
   (( $# )) && value="${first#*=} $*" || value=${first#*=}
-  [[ $value == '('*')' ]] && emit "eval 'declare -a $name=$value'" || emit 'eval "declare -a '"$name"'=${'"$value"'}"'
+  is_literal "$value" && emit "eval 'declare -a $name=$value'" || emit 'eval "declare -a '"$name"'=${'"$value"'}"'
 }
 
 local_hsh () {
-  local first=$1; shift
   local item
-  local key
   local name
   local result=''
   local value
 
-  name=${first%%=*}
-  (( $# )) && value="${first#*=} $*" || value=${first#*=}
-  [[ $value =~ ^[_[:alpha:]][_[:alnum:]]*(\[.+])?$ ]] && {
-    is_set "$value" && value=${!value} || value=''
+  (( $# )) || return
+  name=${1%%=*}
+  value=${1#*=}
+  shift
+  set -- "$value" "$@"
+  case $# in
+    1 )
+      is_set "$value" && value=${!value}
+      shift
+      ;;
+    * ) is_literal "$*" && { value=$*; set -- ;};;
+  esac
+  { [[ -z $value ]] || is_literal "$value" ;} && {
+    ! (( $# )) || return
+    emit "eval 'declare -A $name=$value'"
+    return
   }
-  [[ $value == '('*')' ]] && { emit "eval 'declare -A $name=$value'"; return ;}
-  for item in $value; do
+  for item in "$value" "$@"; do
     [[ $item == *?=* ]] || return
-    key=${item%%=*}
-    result+="[$key]=${item#*=} "
+    printf -v result '%s [%s]=%q' "$result" "${item%%=*}" "${item#*=}"
   done
-  result="( $result )"
-  emit "eval 'declare -A $name=$result'"
+  emit "eval 'declare -A $name=($result )'"
 }
 
 log () { put "$@" ;}
@@ -248,8 +266,8 @@ parse_options () {
   case ${#arg_ary[@]} in
     '0' ) statement='set --';;
     *   )
-      printf -v statement '%q ' "${arg_ary[@]}"
-      printf -v statement 'set -- %s' "$statement"
+      escape_items "${arg_ary[@]}"
+      printf -v statement 'set -- %s' "$__"
       ;;
   esac
   repr result_hsh
@@ -268,6 +286,13 @@ part () {
 
 put     () { printf '%s\n' "$@"   ;}
 puterr  () { put "Error: $1" >&2  ;}
+
+raise () {
+  local rc=$?
+
+  [[ -n ${1:-} ]] && puterr "$1"
+  emit "return ${2:-$rc}"
+}
 
 repr () {
   __=$(declare -p "$1" 2>/dev/null) || return
@@ -313,13 +338,15 @@ require_relative () {
   local extension_ary=()
   local file
 
+  $(with __macros)
+
   extension_ary=(
     .bash
     .sh
     ''
   )
   [[ $spec != /* && $spec == *?/* ]] || return
-  caller_dir=$(readlink -f "$(dirname "$(readlink -f "${BASH_SOURCE[1]}")")")
+  caller_dir=$($readlink "$(dirname "$($readlink "${BASH_SOURCE[1]}")")")
   file=$caller_dir/$spec
   for extension in "${extension_ary[@]}"; do
     [[ -e $file$extension ]] && break
@@ -356,8 +383,8 @@ strict_mode () {
 }
 
 stuff () {
-  [[ $2 == 'into'   ]] || return
-  [[ $1 == '('*')'  ]] && eval "local -a ref_ary=$1" || local -a ref_ary=( "$1" )
+  [[ $2 == 'into' ]]  || return
+  is_literal "$1"     && eval "local -a ref_ary=$1" || local -a ref_ary=( "$1" )
   $(local_hsh result_hsh=$3)
   local ref